restrict item management

+++ Controllers/ItemController.cs
@@ -3,11 +3,13 @@ using LabManAPI.Models;
 using Microsoft.AspNetCore.Mvc;
 using LabManAPI.Contracts;
 using LabManAPI.Services;
-using System;
 using LabManAPI.Contracts.Requests;
 using LabManAPI.Contracts.Responses;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
+using LabManAPI.Extensions;
+using Microsoft.AspNetCore.Authentication;
+using Microsoft.AspNetCore.Http;
 
 namespace LabManAPI.Controllers
 {
@@ -16,20 +18,35 @@ namespace LabManAPI.Controllers
     {
         private readonly IItemService _itemService;
 
-        public ItemController(IItemService itemService)
+        private readonly IGeneralExtensions _extensions;
+
+        public ItemController(IItemService itemService, IGeneralExtensions extensions)
         {
             _itemService = itemService;
+            _extensions = extensions;
         }
 
         [HttpGet(ApiRoutes.Item.GetAll)]
         public async Task<IActionResult> GetAll()
         {
+            var isAdmin = await _extensions.isAdmin(await HttpContext.GetTokenAsync("access_token"));
+            if (!isAdmin)
+            {
+                return StatusCode(StatusCodes.Status403Forbidden, "User is not allowed");
+            }
+
             return Ok(await _itemService.GetItemsAsync());
         }
 
         [HttpPost(ApiRoutes.Item.Create)]
         public async Task<IActionResult> Create([FromBody] CreateItemRequest itemRequest)
         {
+            var isAdmin = await _extensions.isAdmin(await HttpContext.GetTokenAsync("access_token"));
+            if (!isAdmin)
+            {
+                return StatusCode(StatusCodes.Status403Forbidden, "User is not allowed");
+            }
+
             var item = new Item
             {
                 Name = itemRequest.Name,
@@ -57,6 +74,12 @@ namespace LabManAPI.Controllers
         [HttpDelete(ApiRoutes.Item.Delete)]
         public async Task<IActionResult> Delete([FromRoute] int itemId)
         {
+            var isAdmin = await _extensions.isAdmin(await HttpContext.GetTokenAsync("access_token"));
+            if (!isAdmin)
+            {
+                return StatusCode(StatusCodes.Status403Forbidden, "User is not allowed");
+            }
+
             var deleted = await _itemService.DeleteItemAsync(itemId);
 
             if (deleted)
@@ -69,6 +92,11 @@ namespace LabManAPI.Controllers
         [HttpPut(ApiRoutes.Item.Update)]
         public async Task<IActionResult> Update([FromBody] UpdateItemRequest request, [FromRoute] int itemId)
         {
+            var isAdmin = await _extensions.isAdmin(await HttpContext.GetTokenAsync("access_token"));
+            if (!isAdmin)
+            {
+                return StatusCode(StatusCodes.Status403Forbidden, "User is not allowed");
+            }
 
             var item = await _itemService.GetItemByIdAsync(itemId);
